mlops/requirements.txt
qdrant-client==1.7.0
sentence-transformers==2.3.1
huggingface-hub==0.20.0
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
tqdm==4.66.1
